Downgrade JetBrains annotations to 13.0 for compatibility with Alloy2B
ProB 2 uses Alloy2B, which uses Kotlin, which depends on version 13.0
of the JetBrains annotations. These are not fully compatible with newer
versions of the annotations (in particular, @Nullable and @NotNull do
not support TYPE_USE targets).

Because Alloy2B is distributed as an uberjar that contains all of its
dependencies, we cannot override the JetBrains annotations 13.0
dependency without excluding Alloy2B completely. So instead we have to
make our code compatible with the 13.0 annotations (which means
removing annotations in a few places).