Add initial implementation of new command argument parsing
The command arguments are now parsed before Command.run is called,
based on a description of the parameter structure that must be provided
by each command.

For now, only "single" and "remainder" parameters are supported (both
required and optional). There is no native support for repeated
parameters yet.

The parsed argument values can be read by the command implementations
in a type-safe way. For example, a required parameter is returned as a
String and an optional parameter as an Optional<String>, and there are
no generic Object return values that need to be manually casted first.