Remove PositionalParameter (except as a namespace/utility class)
At the moment, positional parameters are the only supported kind of
parameter, so the distinction between parameters in general and
positional parameters is not necessary yet, and the current supposedly
generic Parameter class already contains some aspects specific to
positional parameters (e. g. argument splitting). It makes more sense
to merge PositionalParameter into Parameter for now, and work out a
proper common superclass/interface once support for non-positional
parameters (options/flags) is added.

The PositionalParameter class remains for now, but only as a namespace
for the specific parameter classes (PositionalParameter.RequiredSingle,
etc.) - this avoids having to update each reference to the specific
parameter classes for now. The specific parameter classes now extend
Parameter instead of PositionalParameter, but this does not affect most
usages.