Change kernelspec JSON generation to use JsonElement APIs
This way we do not need to define a custom class with naming
annotations, etc. and can instead create the needed JSON structure
directly by hand. We only need to use this JSON structure exactly once,
so it's not worth the effort to properly map it to a native Java class
representation.