Implement kernel interrupts properly
Jupyter's default interrupt behavior is to send a signal to the kernel
process, which terminates the process unless the process handles it
(which Java does not allow). By changing the kernel.json interrupt_mode
setting to message, Jupyter will instead send a regular message to the
kernel to request an interrupt, which is handled by the basekernel and
passed to the kernel implementation.

Our interrupt implementation now interrupts the Java thread of the
currently running command and sends an interrupt to the current state
space (which interrupts the underlying probcli). This should be
sufficient to interrupt all long-running code in probcli and ProB 2.