Clean up help output of main class (installer)
If the user passes no arguments, the full help is no longer displayed
automatically. Instead, only a short message is displayed that explains
how to install the kernel and how to view the full help. There is now
also more whitespace in the help text to make it more readable.