Re-implement command inspection feature based on new argument parsing
CommandUtils.splitArgs now takes an extra (optional) parameter to ask
it to not split the entire argument string, but only up to the argument
at the given offset in the string. The returned SplitResult contains
information about which parameter the argument splitting stopped at.

This is used in the new implementation of the inspection feature: when
the kernel is asked to inspect at a certain position, the arguments are
split up to that position, and the argument at that position is
inspected. (The arguments are only split and not fully parsed, because
inspection should be possible even if the command arguments are still
incomplete or otherwise invalid.)

This new implementation replaces the old separate implementation in
CommandUtils.splitArgs.