Remove unnecessary special case in inspection/completion
If the cursor is after the command name, but before the start of the
argument string, it makes more sense from a user perspective to treat
it as if it was at the start of the argument string. This way if there
are multiple spaces between the command name and argument, the user can
still use code completion inside this whitespace and get completions
for the first parameter of the command.