Implement error highlighting for code not from the current cell
When an error points to an external file, its source code is now read
and errors highlighted in it.

Error highlighting for code from the last load cell now also works when
an error pointing to it is thrown in a later cell (although this
doesn't happen often). The initial/default machine is also recognized
(although there should never be errors pointing to its code).