Rewrite installation process to be fully Python-based
This allows using Jupyter's Python API to install the kernel. It returns
the location of the newly installed kernelspec, which allows copying the
kernel jar file into the kernelspec directory instead of the ProB home
directory. This is cleaner and allows fully uninstalling the kernel
using a single command.

Additionally, we no longer depend on Guice to generate the kernel.json
(we use Python's json module instead) and no longer use the deprecated
Main.getProBDirectory method from ProB 2.

The Python script for installing the kernel is bundled into the kernel
jar as __main__.py, which allows running the jar file using Python. The
installation can still be started using Java, in which case the jar
re-executes itself using Python. Yes, this is a little bit cursed, but
it works.