Rename Event-B result/dummy variables to also be valid B identifiers
Needed for compatibility with the next SNAPSHOT version of the ProB Java
API, where PredicateBuilder adds backquotes around variable names that
aren't valid B identifiers. This causes syntax errors when building an
Event-B predicate with variable names that are valid in Event-B, but not
in B.

The proper fix would be to add some way to tell PredicateBuilder what
the expected output language is, but the quick and easy fix is to just
change the internal variable names.